[IT-4462] Update to a supported Beanstalk solution stack (#106)
* Update to an AWS supported solution stack for `develop` environment

* Changing the solution stack required destroying and re-creation of the
beanstalk environment which created a new load balancer. Updated the
reference to the new load balancer.

* There is a circular dependency in the Sage-Bionetworks/synapse-login-scipool code
when using the ssm client to get the region info from the SSM. The SSM client needs
the region of the SSM before it can be used to lookup a value from the SSM.
Instead of getting the region info from the SSM we change the AWS_REGION ENV
variable to contain the direct region value.